Cache placeholder definitions for components to be harvested
Over a 24-hour period, statistics show that half of the coordinates with an unavailable definition are requested multiple times. To optimize for such cases, we will cache the placeholder definition computed for components that are pending harvest. Future requests for the same coordinates will utilize these cached placeholders, improving performance and preventing duplicate harvest requests.

The placeholder definition
- expires in 1 days (half the time for available or recomputed definitions), if the component is not harvested.
- will be replaced by the computed defintion when the tool results for component is available.